backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
numpy==1.26.2
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
python-jose==3.3.0
python-dotenv==1.0.0
pydantic==2.5.1